Pull up r19052:r19183 from trunk
git-svn-id: svn://anonsvn.mit.edu/krb5/users/coffman/pkinit@19184 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/ktdefname.c
@@ -37,18 +37,18 @@ extern char *krb5_defkeyname;
 char *krb5_overridekeyname = NULL;
 
 krb5_error_code KRB5_CALLCONV
-krb5_kt_default_name(krb5_context context, char *name, int namesize)
+krb5_kt_default_name(krb5_context context, char *name, size_t namesize)
 {
     char *cp = 0;
     char *retval;
 
     if (krb5_overridekeyname) {
-	if ((size_t) namesize < (strlen(krb5_overridekeyname)+1))
+	if (namesize < (strlen(krb5_overridekeyname)+1))
 	    return KRB5_CONFIG_NOTENUFSPACE;
 	strcpy(name, krb5_overridekeyname);
     } else if ((context->profile_secure == FALSE) &&
 	(cp = getenv("KRB5_KTNAME"))) {
-	if ((size_t) namesize < (strlen(cp)+1))
+	if (namesize < (strlen(cp)+1))
 	    return KRB5_CONFIG_NOTENUFSPACE;
 	strcpy(name, cp);
     } else if ((profile_get_string(context->profile,
@@ -56,7 +56,7 @@ krb5_kt_default_name(krb5_context context, char *name, int namesize)
 				   "default_keytab_name", NULL, 
 				   NULL, &retval) == 0) &&
 	       retval) {
-	if ((size_t) namesize < (strlen(retval)+1))
+	if (namesize < (strlen(retval)+1))
 	    return KRB5_CONFIG_NOTENUFSPACE;
 	strcpy(name, retval);
 	profile_release_string(retval);
@@ -73,7 +73,7 @@ krb5_kt_default_name(krb5_context context, char *name, int namesize)
 	    sprintf(name, krb5_defkeyname, defname);
 	}
 #else
-	if ((size_t) namesize < (strlen(krb5_defkeyname)+1))
+	if (namesize < (strlen(krb5_defkeyname)+1))
 	    return KRB5_CONFIG_NOTENUFSPACE;
 	strcpy(name, krb5_defkeyname);
 #endif
